moved components closer to where they are being used
