+++ src/app/(pokemon)/stats/Stats.tsx
